requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
nltk
transformers
datasets
scipy
torch
kafka-python
fastapi
uvicorn[standard]
jinja2
python-multipart
pymongo
elasticsearch==8.7.0